Populate the gguid isPermaLink field which is required by some 3rd parties.
"guid must be a full URL, unless isPermaLink attribute is false" (https://validator.w3.org/feed/docs/error/InvalidPermalink.html)

In late 2018 we started using internal ids for the guid. These are not full urls so we set to false after this date.

One of the items in the test file has had it's date shifted to exercise this case.