Use the shorter content trailtext to populate the itunes:subtitle tag rather than standfirst. Limit to 255 characters.
Drop back to an empty (but present) tag if the trailtext is over length. There is already precedent in the code for this approach.

Reacts to w3c feed validation errors. The original specification for itunes:subtitle is lost to history and we can't find an up to date citation other than the w3c feed validator.